+++ NackUn/StudyFirst/app/src/main/java/sample/nackun/com/studyfirst/vo/Market.kt
@@ -1,4 +1,4 @@
-package sample.nackun.com.studyfirst.market
+package sample.nackun.com.studyfirst.vo
 
 data class Market(
     val market: String
